@@ -22,36 +22,38 @@
   },
   "devDependencies": {
     "@iconify-json/mdi": "^1.2.3",
-    "@octokit/plugin-retry": "^7.1.4",
-    "@octokit/plugin-throttling": "^9.4.0",
+    "@octokit/plugin-retry": "^7.2.1",
+    "@octokit/plugin-throttling": "^10.0.0",
     "@octokit/rest": "^21.1.1",
     "@popperjs/core": "^2.11.7",
     "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tailwindcss/vite": "^4.1.7",
     "@tauri-apps/cli": "^2.5.0",
     "@testing-library/svelte": "^5.2.7",
     "@tsconfig/svelte": "^5.0.4",
-    "@vitest/coverage-v8": "^3.0.8",
+    "@vitest/coverage-v8": "^3.1.3",
     "ansi-to-span": "^0.0.1",
     "autoprefixer": "^10.4.21",
     "escape-html": "^1.0.3",
     "flowbite": "^3.1.2",
-    "flowbite-svelte": "0.48.4",
-    "jsdom": "^26.0.0",
+    "flowbite-svelte": "1.2.2",
+    "jsdom": "^26.1.0",
     "postcss": "^8.5.3",
     "postcss-load-config": "^6.0.1",
     "prettier": "^3.5.3",
-    "prettier-plugin-svelte": "^3.3.3",
-    "svelte": "^5.22.6",
+    "prettier-plugin-svelte": "^3.4.0",
+    "svelte": "^5.30.1",
     "svelte-i18n": "^4.0.1",
     "svelte-navigator": "^3.2.2",
     "svelte-preprocess": "^6.0.3",
-    "tailwindcss": "^3.4.13",
-    "typescript": "^5.8.2",
+    "tailwindcss": "^4.1.7",
+    "typescript": "^5.8.3",
     "unplugin-icons": "^22.1.0",
-    "vite": "^6.2.1",
-    "vitest": "^3.0.8"
+    "vite": "^6.3.5",
+    "vitest": "^3.1.3"
   },
   "dependencies": {
+    "@tailwindcss/postcss": "^4.1.7",
     "@tauri-apps/api": "^2.5.0",
     "@tauri-apps/plugin-clipboard-manager": "^2.2.2",
     "@tauri-apps/plugin-dialog": "~2",
