@@ -115,7 +115,7 @@
   <div class="flex flex-row gap-2">
     <Button
       class="border-solid border-2 border-slate-900 rounded bg-slate-900 hover:bg-slate-800 text-sm text-white font-semibold px-5 py-2"
-      on:click={async () => {
+      onclick={async () => {
         launchGame($activeGame, false);
       }}>{$_("gameControls_button_play")}</Button
     >
@@ -123,10 +123,15 @@
       class="text-center font-semibold focus:ring-0 focus:outline-none inline-flex items-center justify-center px-2 py-2 text-sm text-white border-solid border-2 border-slate-900 rounded bg-slate-900 hover:bg-slate-800"
       >{$_("gameControls_button_features")}</Button
     >
-    <Dropdown trigger="hover" placement="top-end" class="!bg-slate-900">
+    <Dropdown
+      simple
+      trigger="hover"
+      placement="top-end"
+      class="!bg-slate-900 dark:text-white **:w-full"
+    >
       <DropdownItem
         disabled={!textureSupportEnabled}
-        on:click={async () => {
+        onclick={async () => {
           navigate(`/${$activeGame}/texture_packs`);
         }}
       >
@@ -137,7 +142,8 @@
         >
       {/if}
       <DropdownItem
-        on:click={async () => {
+        class=""
+        onclick={async () => {
           navigate(`/${$activeGame}/mods`);
         }}
       >
@@ -145,54 +151,59 @@
       </DropdownItem>
     </Dropdown>
     <Button
-      class="text-center font-semibold focus:ring-0 focus:outline-none inline-flex items-center justify-center px-2 py-2 text-sm text-white border-solid border-2 border-slate-900 rounded bg-slate-900 hover:bg-slate-800"
+      class="text-center font-semibold focus:ring-0 focus:outline-none inline-flex items-center px-2 py-2 text-sm text-white border-solid border-2 border-slate-900 rounded bg-slate-900 hover:bg-slate-800"
     >
       {$_("gameControls_button_advanced")}
     </Button>
-    <Dropdown trigger="hover" placement="top-end" class="!bg-slate-900">
+    <Dropdown
+      simple
+      trigger="hover"
+      placement="top-end"
+      class="!bg-slate-900 dark:text-white **:w-full"
+    >
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           launchGame($activeGame, true);
         }}>{$_("gameControls_button_playInDebug")}</DropdownItem
       >
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           launchGameWithCustomExecutable($activeGame);
         }}>Launch with Custom Executable</DropdownItem
       >
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           openREPL($activeGame);
         }}>{$_("gameControls_button_openREPL")}</DropdownItem
       >
       <DropdownDivider />
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           dispatch("job", {
             type: "decompile",
           });
         }}
         >{$_("gameControls_button_decompile")}
         <!-- NOTE - this is a bug in flowbite-svelte, it's not replacing the default class but just appending -->
-        <Helper helperClass="!text-neutral-400 !text-xs"
+        <Helper class="!text-neutral-400 !text-xs"
           >{$_("gameControls_button_decompile_helpText")}</Helper
         ></DropdownItem
       >
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           dispatch("job", {
             type: "compile",
           });
         }}
         >{$_("gameControls_button_compile")}
         <!-- NOTE - this is a bug in flowbite-svelte, it's not replacing the default class but just appending -->
-        <Helper helperClass="!text-neutral-400 !text-xs"
+        <Helper class="!text-neutral-400 !text-xs"
           >{$_("gameControls_button_compile_helpText")}
         </Helper></DropdownItem
       >
       <DropdownDivider />
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           if (gameDataDir) {
             await openDir(gameDataDir);
           }
@@ -204,17 +215,22 @@
     >
       <IconCog />
     </Button>
-    <Dropdown trigger="hover" placement="top-end" class="!bg-slate-900">
+    <Dropdown
+      simple
+      trigger="hover"
+      placement="top-end"
+      class="!bg-slate-900 dark:text-white **:w-full"
+    >
       <!-- TODO - screenshot folder? how do we even configure where those go? -->
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           if (settingsDir) {
             await openDir(settingsDir);
           }
         }}>{$_("gameControls_button_openSettingsFolder")}</DropdownItem
       >
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           if (savesDir) {
             await openDir(savesDir);
           }
@@ -222,25 +238,25 @@
       >
       <DropdownDivider />
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           const launchString = await getLaunchGameString($activeGame);
           await writeText(launchString);
           toastStore.makeToast($_("toasts_copiedToClipboard"), "info");
         }}
         >{$_("gameControls_button_copyExecutableCommand")}<Helper
-          helperClass="!text-neutral-400 !text-xs"
+          class="!text-neutral-400 !text-xs"
           >{$_("gameControls_button_copyExecutableCommand_helpText_1")}<br
           />{$_("gameControls_button_copyExecutableCommand_helpText_2")}</Helper
         ></DropdownItem
       >
       <DropdownDivider />
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           await resetGameSettings($activeGame);
         }}>{$_("gameControls_button_resetSettings")}</DropdownItem
       >
       <DropdownItem
-        on:click={async () => {
+        onclick={async () => {
           // Get confirmation
           // TODO - probably move these confirms into the actual launcher itself
           const confirmed = await confirm(
@@ -253,7 +269,7 @@
           }
         }}
         >{$_("gameControls_button_uninstall")}<Helper
-          helperClass="!text-neutral-400 !text-xs"
+          class="!text-neutral-400 !text-xs"
           >{$_("gameControls_button_uninstall_helpText")}</Helper
         ></DropdownItem
       >
