@@ -146,7 +146,7 @@
       <div class="flex flex-row gap-2">
         <Button
           class="border-solid border-2 border-slate-900 rounded bg-slate-900 hover:bg-slate-800 text-sm text-white font-semibold px-5 py-2"
-          on:click={async () => dispatch("change")}
+          onclick={async () => dispatch("change")}
           >{$_("setup_button_continue")}</Button
         >
       </div>
@@ -166,7 +166,7 @@
         <!-- TODO - no button to go back -->
         <Button
           class="border-solid border-2 border-slate-900 rounded bg-slate-900 hover:bg-slate-800 text-sm text-white font-semibold px-5 py-2"
-          on:click={async () => await generateSupportPackage()}
+          onclick={async () => await generateSupportPackage()}
           >{$_("setup_button_getSupportPackage")}</Button
         >
       </div>
@@ -182,13 +182,13 @@
     <div class="flex flex-row gap-2">
       <Button
         class="border-solid border-2 border-slate-900 rounded bg-slate-900 hover:bg-slate-800 text-sm text-white font-semibold px-5 py-2"
-        on:click={async () => await install(false)}
+        onclick={async () => await install(false)}
         >{$_("setup_button_installViaISO")}</Button
       >
       <!-- TODO - disabled for now, needs fixes in the extractor -->
       <!-- <Button
         class="border-solid border-2 border-slate-900 rounded bg-slate-900 hover:bg-slate-800 text-sm text-white font-semibold px-5 py-2"
-        on:click={async () => await install(true)}
+        onclick={async () => await install(true)}
         >Install via Folder</Button
       > -->
     </div>
