@@ -25,21 +25,21 @@
     <Button
       disabled={downloadingPackage}
       class="border-solid rounded bg-orange-400 hover:bg-orange-600 text-sm text-slate-900 font-semibold px-4 py-2"
-      on:click={async () => {
+      onclick={async () => {
         downloadingPackage = true;
         await generateSupportPackage();
         downloadingPackage = false;
       }}
     >
       {#if downloadingPackage}
-        <Spinner class="text-sm mb-0.5 mr-1" size="4" color="white" />
+        <Spinner class="text-sm mb-0.5 mr-1" size="4" color="yellow" />
       {/if}
       {$_("help_button_downloadPackage")}</Button
     >
     {#if appLogDirPath}
       <Button
         class="flex items-center border-solid rounded bg-white hover:bg-orange-400 text-sm text-slate-900 font-semibold px-4 py-2"
-        on:click={() => {
+        onclick={() => {
           openDir(appLogDirPath);
         }}>{$_("help_button_openLogFolder")}</Button
       >
