@@ -48,23 +48,23 @@
     <div class="flex flex-row mt-1 gap-3">
       <Button
         class="border-solid rounded bg-orange-400 hover:bg-orange-600 text-sm text-slate-900 font-semibold px-5 py-2"
-        on:click={async () => await updateHandler()}
+        onclick={async () => await updateHandler()}
         disabled={updating}
       >
         {#if updating}
-          <Spinner class="mr-3" size="4" color="white" />
+          <Spinner class="mr-3" size="4" color="yellow" />
         {/if}
         {$_("update_button_doUpdate")}
       </Button>
       <Button
         class="flex-shrink border-solid rounded bg-white hover:bg-orange-400 text-sm text-slate-900 font-semibold px-5 py-2"
-        on:click={() => (showChanges = !showChanges)}
+        onclick={() => (showChanges = !showChanges)}
         >{$_("update_button_viewChangelog")}</Button
       >
       <Toggle
         checked={showDependencyChanges}
         color="orange"
-        on:change={(evt) => {
+        onchange={(evt) => {
           showDependencyChanges = evt.target.checked;
         }}>{$_("update_button_hideDependencyChanges")}</Toggle
       >
